requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
audioread==3.0.1
av==12.3.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cmake==3.30.2
coloredlogs==15.0.1
cryptography==43.0.1
ctranslate2==4.3.1
decorator==5.1.1
diskcache==5.6.3
exceptiongroup==1.2.2
fastapi==0.114.0
faster-whisper==1.0.3
filelock==3.15.4
flatbuffers==24.3.25
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.8
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
narwhals==1.6.2
nexaai==0.0.8.2
numba==0.60.0
numpy==2.0.2
onnxruntime==1.19.2
packaging==24.1
pandas==2.2.2
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
prompt_toolkit==3.0.47
protobuf==5.28.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.0
pydantic_core==2.23.2
pydeck==0.9.1
Pygments==2.18.0
pypdfium2==4.30.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.8.0
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.1
scipy==1.14.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
starlette==0.38.4
streamlit==1.38.0
streamlit-audiorec==0.1.3
sympy==1.13.2
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
wcwidth==0.2.13
